move gaps-specific logic out of con.c and render.c into gaps.c
